All the code has been restructured.
At this point, it will scrape odds from Neds and SportsBet, taking the team names, odds, and betting agency and putting all of that data into a Game object. There are two things that I need to do now: 1. add more betting agencies, and 2. work out how to find out when different sites are offering odds on the same game so that I can compare them.